requirements.txt
unstructured==0.12.6
tiktoken==0.6.0
pinecone-client==3.1.0
pypdf==4.1.0
openai==1.14.2
langchain==0.1.13
pandas==2.2.1
numpy==1.26.4
jupyter==1.0.0
ipykernel==6.29.3
langchain-openai==0.1.0
Flask==3.0.2
python_dotenv==1.0.1
langchain-community
flask_cors==4.0.0
langchain_pinecone==0.0.3